--- test/support/resampling/ResamplingHelpers.h
@@ -1,15 +1,0 @@
-#pragma once
-
-// Umbrella include for resampling behavioral tests and calibration probes.
-// Layers: processors → signal → chain → measurements → contract assertions → fixtures.
-
-#include "TestInfrastructure.h"
-#include "HostConfigCatalog.h"
-#include "PassthroughProcessor.h"
-#include "DelayLineProcessor.h"
-#include "SimulatedOnnxProcessor.h"
-#include "ResamplingSignalUtils.h"
-#include "ResamplingChainHelpers.h"
-#include "ResamplingMeasurements.h"
-#include "ResamplingContractAssertions.h"
-#include "ResamplingFixtures.h"
